requirements.txt
python-dotenv
faker 
psycopg2-binary
